refactor: Improved project structure
