+++ internal/queue/topic.go
@@ -10,27 +10,6 @@ import (
 	"time"
 )
 
-type Queue[T any] interface {
-	Enqueue(T) int
-	Dequeue() (T, error)
-	Size() int
-	Cap() int
-}
-
-type TopicConfig struct {
-	AckTimeout time.Duration
-	MaxRetries int
-}
-
-// Message is a simple struct holding the message and data
-type Message struct {
-	ID        int
-	Payload   string
-	Timestamp time.Time // When it was delivered
-	Acked     bool      // Whether it's been acknowledged
-	Retries   int
-}
-
 // Topic represents a named message queue
 type Topic struct {
 	Name     string
